+++ server/ntb/commands/sync_topics/__init__.py
@@ -1,0 +1,1 @@
+from .command import SyncTopicsCommand  # NOQA
