+++ server/ntb/commands/sync_topics_reports/2021-08-20T16-14.md
@@ -1,0 +1,33 @@
+
+New Items:
+----------
+No new items discovered
+
+Missing Norwegian Translation:
+------------------------------
+All Topics are translated
+
+Deviated Items:
+---------------
+
+| Qcode | Field | Existing | IPTC |
+| ----- | ----- | -------- | ---- |
+| 20001179 | wikidata | Q20937557 | Q7725310 |
+| 20001200 | parent | 20000118 | 20000131 |
+| 20000175 | parent | 20000171 | 20000183 |
+| 20000184 | parent | 20000171 | 20000183 |
+| 20000186 | parent | 20000171 | 20000183 |
+| 20000179 | parent | 20000178 | 20000171 |
+| 20001167 | wikidata | Q6657015 | Q63383285 |
+| 20000330 | wikidata | Q1566 | Q907703 |
+| 20000498 | parent | 20000497 | 08000000 |
+| 20001238 | parent | 20000502 | 08000000 |
+| 20000505 | parent | 20000504 | 08000000 |
+| 20000504 | parent | 20000502 | 08000000 |
+| 20000503 | parent | 20000502 | 08000000 |
+| 20000499 | parent | 20000497 | 08000000 |
+| 20000608 | wikidata | Q274490 | Q1929688 |
+| 20000818 | wikidata | Q456673 | Q1077064 |
+| 20001240 | parent | 15000000 | 20000822 |
+| 20000887 | parent | 20000884 | 20000822 |
+| 20001148 | wikidata | Q56072650 | Q18515440 |
